[process][signal] Move to sys_rt_sigreturn() to the portable code in signal.c
