+++ kernel/signal.c
@@ -12,6 +12,7 @@
 #include <tilck/kernel/sys_types.h>
 #include <tilck/kernel/hal.h>
 #include <tilck/kernel/interrupts.h>
+#include <tilck/kernel/process_int.h>
 
 #include <tilck/mods/tracing.h>
 
@@ -851,4 +852,57 @@ int sys_sigaction(ulong a1, ulong a2, ulong a3)
 __sighandler_t sys_signal(int signum, __sighandler_t handler)
 {
    NOT_IMPLEMENTED(); // deprecated interface
+}
+
+ulong sys_rt_sigreturn(void)
+{
+   ASSERT(!is_preemption_enabled()); /* Thanks to SYSFL_NO_PREEMPT */
+   struct task *curr = get_curr_task();
+   regs_t *r = curr->state_regs;
+   ulong user_sp;
+
+   if (LIKELY(curr->nested_sig_handlers > 0)) {
+
+      trace_printk(10, "Done running signal handler");
+
+      user_sp = regs_get_usersp(r);
+      user_sp +=
+         sizeof(ulong)               /* compensate the "push signum" above    */
+         + SIG_HANDLER_ALIGN_ADJUST; /* compensate the forced stack alignment */
+      regs_set_usersp(r, user_sp);
+
+      if (!process_signals(curr, sig_in_return, r)) {
+
+         if (curr->in_sigsuspend) {
+            memcpy(curr->sa_mask, curr->sa_old_mask, sizeof(curr->sa_mask));
+            curr->in_sigsuspend = false;
+         }
+
+         restore_regs_from_user_stack(r);
+      }
+
+      curr->nested_sig_handlers--;
+      ASSERT(curr->nested_sig_handlers >= 0);
+
+   } else {
+
+      /* An user process tried to call directly rt_sigreturn() */
+      set_return_register(r, (ulong) -ENOSYS);
+   }
+
+   /*
+    * NOTE: we must return r->eax because syscalls are called by handle_syscall
+    * in a generic way like:
+    *
+    *     r->eax = (ulong) fptr(...)
+    *
+    * Returning anything else than r->eax would change that register and we
+    * don't wanna do that in a special NORETURN function such this. Here we're
+    * supposed to restore all the user registers as they were before the signal
+    * handler ran. Failing to do that, has an especially visible effect when
+    * a signal handler run after preempting running code in userspace: in that
+    * case, no syscall was made and no register is expected to ever change,
+    * exactly like in context switch.
+    */
+   return get_return_register(r);
 }
